add JavaDoc

+++ ProjetProxiBanqueV1/src/metier/CompteEntreprise.java
@@ -4,14 +4,16 @@ package metier;
 public class CompteEntreprise extends CompteCourant {
 	
 	//constructeur
-		public CompteEntreprise(int numCompte, float solde, String dateOuverture, double decouvert) {
+		public CompteEntreprise(int numCompte, double solde, String dateOuverture, double decouvert) {
 			super(numCompte, solde, dateOuverture);
 			this.decouvert = decouvert;
 		}
-		public CompteEntreprise(int numCompte, float solde, String dateOuverture) {
+		public CompteEntreprise(int numCompte, double solde, String dateOuverture) {
 			super(numCompte, solde, dateOuverture);
 			this.decouvert = 1000.00;
 			
 		}
+		
+		
 		
 }
